requirements.txt
absl-py==1.4.0
beautifulsoup4==4.12.2
boxmot==10.0.42
cachetools==5.3.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
distlib==0.3.7
filelock==3.12.2
filterpy==1.4.5
fonttools==4.42.1
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
identify==2.5.27
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lapx==0.5.4
loguru==0.7.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
nodeenv==1.8.0
numpy==1.23.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.1
pika==1.3.2
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==3.3.3
protobuf==4.24.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.160
urllib3==1.26.16
virtualenv==20.24.4
wcwidth==0.2.6
Werkzeug==2.3.7
win32-setctime==1.1.0
yacs==0.1.8
pyyaml